+++ src/summary.h
@@ -1,3 +1,27 @@
+/* budgurse - budgeting with curses
+ *
+ * Copyright (c) 2025 W. C. Aitken 
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to
+ * deal in the Software without restriction, including without limitation the
+ * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
+ * sell copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions: 
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software. 
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+ * IN THE SOFTWARE.
+ *
+ */
+
 #ifndef SUMMARY_H
 #define SUMMARY_H
 
@@ -6,6 +30,7 @@
 #include "wins.h"
 #include "browser.h"
 #include "util.h"
+#include "status.h"
 #include "global.h"
 #include "budgurse.h"
 
@@ -34,7 +59,7 @@ void free_summary(summary_t* s);
 
 void summary_calc();
 int summary_update_on_entry(entry_t *e);
-void summary_inc_cell(int x, int y, int value);
+void summary_inc_cell(int x, int y, float value);
 int summary_set_date_bounds(time_t *n_date, time_t *o_date, delin_t d);
 void summary_clear();
 void summary_reset(time_t max_date, delin_t d);
